lowered stork speed (game was too hard with stork flying 1 blok per second), widened the distance from which frog can jump on car.

+++ zabcia/config.h
@@ -14,7 +14,7 @@ typedef struct {
     int bonuses;
     int bonus_time;
     int stork_spawn;
-    int stork_speed;
+    double stork_speed;
 }LevelConfig;
 
 void read_lvl_config(FILE* file, int level, LevelConfig* config);
